move privacyLevels into components folder
