+++ pages/[user]/components/collections/CollectionFolder.tsx
@@ -10,8 +10,8 @@ import {
 import Link from "next/link";
 
 import CollectionCard from "./CollectionCard";
-import PRIVACY_LEVELS from "./privacyLevels";
 
+import PRIVACY_LEVELS from "components/collections/privacyLevels";
 import { COLLECTION_ICONS } from "components/collections/CollectionIcons";
 import { CardCollection, CollectedCard } from "types/makotools";
 import ResponsiveGrid from "components/core/ResponsiveGrid";
