+++ pages/[user]/components/collections/EditCollection.tsx
@@ -29,8 +29,8 @@ import { useEffect, useState } from "react";
 import { GridContextProvider, GridDropZone, GridItem } from "react-grid-drag";
 
 import EditCollectionCard from "./EditCollectionCard";
-import PRIVACY_LEVELS from "./privacyLevels";
 
+import PRIVACY_LEVELS from "./components/collections/privacyLevels";
 import { COLLECTION_ICONS } from "components/collections/CollectionIcons";
 import { useDayjs } from "services/libraries/dayjs";
 import { CardCollection, CollectedCard } from "types/makotools";
